pmdaproc: fix psargs string for processes with large PIDs
The proc fetch routine was assuming that there are only 6 digits
of PID at the start of the name field - this can be off when the
process PID is larger than normal, resulting in extra whitespace
at the start of the psargs string.

Update test 1893 to check for this condition.