pmdasockets: add checking to the pmStore filter strings
Allow a subset of the ss language so as to ensure special
characters - that might be interpreted by the shell which
ss runs in - are not allowed to be set.  Everything that
could be achieved previously can still be achieved though
since alphabetic equivalents exist for all tokens made up
using special characters.