__pmLogInDom_io is gone
I never was happy with this, and have decided to merge the newer
__pmLogInDom_io with the older __pmLogInDom.

So now there is just one data structure for all internal indom handling
and the fine-grained allocation status from __pmLogInDom_io is now in
__pmLogInDom, allowing for the awful "allinbuf" to be dropped.

Some internal API changes (names and data types) are also included
as a consequence.

libpcp_web is still to come ... there are other changes pending there.