src/libpcp/src/check-statics: convert to using readelf if available
With the increasing use of -flto= in the build, the LTO option leaves
our .o files with symbol tables than nm(1) cannot read correctly,
and in in particular the local static symbols are not reported ... and
this is EXACTLY what check-statics needs.

Fortunately readelf(1) still works, so convert the script to use this
when it is available, else fallback to nm(1).