qa/admin/*package*: add packages for the Slurm Perl module
We have a version of the Slurm Perl module that is packaged and
shipped in pcp-testsuite, but no one seems to use it?

qa/1031 fails unless the Slurm Perl module is installed on one of the
directories in the default PERL5LIB path, similarly the PMDA won't
run without a proper Slurm Perl module install.

As the Slurm Prel  module is now packaged and available in many
distros, this change will encourage it to be installed from the
official repos and qa/1031 will get more coverage.