libpcp_import: support for an archive-append mode
In order to implement sysstat/sar semantics, where an
archive is appended to by short-lived processes - and
not by a long-running process ala pmlogger - we're in
need of an append-mode from the LOGIMPORT(3) API.

This is a prototype of the sorts of changes I believe
we'll need.  I've updated just the one QA test utility
to exercise it at this stage and the rest will follow
once we've settled on appropriate APIs.