configure et al: improved detection of backtrace() and associated hooks
backtrace() is not always in glibc, e.g. OpenBSD 7.0 (where backtrace()
exists in libexecinfo, but unfortunately is broken!).

Also, improved reporting of backtraces (especially when the dynamic
symbols table of an executable is sparse or stripped) requires knowledge
of the start and end addresses of the executable's text segment ...
the required symbols do not have consistent names and may not be
present, depending on the whim of your local ld(1).  So we need more
configure glue to sort this out.