libpcp et al: remove repeated functions using backtrace()
Combine all the mybacktrace() and __pmDumpStack() functions into
a single __pmDumpStack() implementation for glibc-based procedure
call backtracing.

ABI change to __pmDumpStack() so that the previous FILE *f argument
is dropped and output goes to stderr (the only FILE * that was used
across the callers) ... this allows __pmDumpStack() to be used with
atexit() as needed.