src/pmlogger/pmlogger_daily_report.sh: uncompress to temp archive if needed
If any component file of the archive that is passed to
src/pmlogger/pmlogger_daily_report (or yesterday's one that it uses
by default) has been compressed, then try to uncompress the archive
before running all the pmrep commands.

Avoids repeated (26 times in the current script) decompression.

Also tidy up the test for an existing archive that works for any
possible compression tool (was just testing for .index or .index.xz),
now tests for .meta*).