qa/1130: check Python version, extend sleeps
Thanks to Nathan (informed guess) and Andreas (detailed analysis)
we now know that the current version of pcp2elasticsearch requires
Python 3.3 or later ... so this is now checked (and we're _notrun on
vm29 RHEL Server 7.9 now instead of failing).

Also noticed a small timing issue ... sleep 1 is rarely robust and
I tripped over the socat being killed before the expected output had
been seen ... changed all of these to sleep 2.