libpcp_web: run pmFetchArchive() in a worker thread
Run pmFetchArchive() in a worker thread instead of the main thread
because it uses blocking (synchronous) I/O calls.

This way the libuv loop does not get blocked, and queued Redis requests
can be sent out while the worker thread is reading PCP archives. As a
side effect, this also breaks the recursion as mentioned in #1605.

Related: #1605