qa/check: add support for "triaged" tests
"triaged" tests are ones where failure is either:
(a) the result of some non-determinism in the test on a specific
    platform such that it sometimes fails, e.g. due to load or
    concurrent activity or collision with cron or systemd timer events
    and this failure is recurring but not repeatable (i.e. the test
    usually passes when rerun), or
(b) something specific to a platform that is not a PCP problem and
    not something we can fix or or filter away or workaround or
    _notrun, or
(c) something that is a known issue, a fix is planned or in progress,
    but is not going to be done for a while.

In all cases, marking the test as "failed" is a bit extreme as we've
already invested effort in understanding the problem, so "triaged" is
a softer failure so we can focus on the other failures that are more
likely to be regressions or test implementation that needs improvement.

See the qa/triaged file for an explanation of what's required to mark
a test as "triaged" on one or more platforms.