qa/common.secure: if no secure_sockets in libpcp, then notrun
With the change to OpenSSL, we're seeing a number of the older platforms
where we are losing the secure_sockets capability ... this is causing
multiple QA failures ... making them _notun seems the best option at
this point.

Most recently this was an issue on vm27 (Ubuntu 16.04) although it
is also happening on vm14 (CentOS6.10), vm29 (RHEL Server 7.9) and
vm33 (OpenBSD 7.0).