libpcp_web: resolve a series context load path end race
There was a path at the completion of successfully loading
an archive via pmseries --load where we cleared the error
condition (PM_ERR_EOL) in a way that raced with a thread
still needing to check for that condition.  End result was
an intermittent segv in the time window checks for archive
completion.

Relates to https://github.com/performancecopilot/pcp/pull/1634