qa/triaged & qa/878: change in triaged file format
qa/878 has always been a little touchy, and had some explicit hosts
(vm01 and bozo-vm) checked for and _notrun in the test itself ... when
it failed the same way on another host (vm29) I decided to move this
into the "triaged" class.

At the same time, the triaged file is getting hard to read, mainly
because the <comments> are (rightly) becoming quite long to explain why
the test is marked "triaged" ... so allow the <comments> to flow over
multiple lines as long as the subsequent lines begin with white space.