libpcp_web: ensure a valid baton is passed to query functions
In the situation where we are calling a function based on an
expanded fabricted SID, we need to pass the userdata rather
than the baton pointer, as this is holding the baton in this
special case situation.  If not we may attempt dereferencing
an invalid pointer in the high level API calling code, such
as pmproxy, when handling error conditions (pmseries_log).

This resolves a crash when handling query division, but logic
errors remain to be sorted out on that code path (fix is WIP).

Related to https://github.com/performancecopilot/pcp/issues/1636