qa/1251: change from pmconfirm to pmdatxmon
This test is checking man page entries and indexing, but since some
older platforms no longer have a version Qt we can play with, then
pmconfirm is not being built and the man pages not installed.

So change from {pmconfirm,pmmessage,pmquery} to
{pmdatxmon,txrecord,genload} which shares exactly the same troff
source code structure and the latter is in the "pcp" package and so
will always be installed.