src/pmlogger/pmlogger_check.sh: improve -V diagnostics around pmlogconf
On bozo pmlogconf was not being run on config.default after some
zeroconf changes (this was unexpected), and the existing -V or -VV
diagnostics gave no clue as to why.

The additional diagnositics helped identify the issue, namely
config.default had been manually updated with pmlogconf at some point,
and the

    Auto-generated by pmlogconf on: <date>

line was missing, so pmlogger_check (rightly) did not dink with it.