src/pmlogconf: additional config files for pmlogger_daily_report
Now that pmlogger_daily_report is back in the pool (at least for those
with the zeroconf package installed), some investigation has revealed
that the pmrep executions use by pmlogger_daily_report require some
metrics that are not included in the default pmlogger configuration
config.default, as generated by pmlogconf -c or re-generated by
pmlogconf -r -c.

So, added ...
    disk.all.discard
    disk.all.blkdiscard
to disk/summary as these seemed generally useful in addition to being
required for pmlogger_daily_report.

And added new zeroconf/... files for the additional metrics.
    src/pmlogconf/zeroconf/disk
    src/pmlogconf/zeroconf/filesystem
    src/pmlogconf/zeroconf/numa
    src/pmlogconf/zeroconf/rpc
    src/pmlogconf/zeroconf/tty

These pmlogconf configurations are all marked with an "include"
control line so they are unconditional, and are either
(a) metrics that also appear as "available" in other configurations ...
    but these ones will NOT be included when pmlogconf is not run
    interactively (as is the case when it is run from pmlogger_check)
(b) metrics that were not in any other pmlogconf configuration