libpcp_web: Redis proxying: support Redis pipelining
The buffer can contain multiple Redis requests (Redis pipelining).

Previously, if the buffer contained multiple Redis requests,
redisSlotsRequest was called with multiple Redis requests.
This triggered an assertion in hiredis code, because only one reply
per request is expected, but in this case it received multiple replies.

This fixes qa/1214.