src/pmlogger/pmlogger_check.sh: small diagnostic change
When trying to match the <pid> of a runing pmlogger, a recent change
added some ps(1) output when -VV is used ... this busted the "... try
<port> host=... match=<n>" line by pasting the ps(1) heading on the
end of that line, and splitting the result text "different directory,
skip", "pmlogger process 3069052 identified, OK", etc the onto a
new line.

This caused QA fallout as well as being hard to read!

Re-order things so the result text is appended correctly, and then
do the ps(1) thing, but only if we're running with -VVV (very, very,
verbose) so the capability is preserved without damaging QA.