build: use fixed copies of pcp and pcpqa sysusers files
For reasons discussed in #1624 generating these files is
not really as helpful as first thought.  Fortunately it
turns out the users and paths therein are the same for us
across Linux distros, so generating them is a moot point.

So, instead, keep a copy of these two one-line files in
the git tree.  Since we keep other files with embedded
hard-coded paths below debian/ and the rpm build picks
those other files up from there already, I've put these
in the same location for now.