build: rpm buildsubdir macro not the right thing to use
The exact reason escapes me, because we use this in a few
other places, but for %pre macros its evidently bad news;
ultimately it fails to run the useradd commands until too
late (after file installations) with this.

Sync'd up a few missing 'exit 0' calls between rpm specs
as well - unclear which is truly correct so play it safe.