pmdasmart: set LC_ALL=C before running smartctl
The smartctl output uses a dot as a thousands separator when the system
is configured with a German locale. The input parsing code didn't handle
this case correctly.

This commit runs LC_ALL=C smartctl, i.e. resetting the locale to a
fixed locale, irrespective of the user's configured locale.