src/pmlogger/pmlogger_daily.sh: break one systemd dependency
We don't need to check pmlogger_daily_report.timer here ... it
is simpler to see if the zeroconf pmlogconf files (or at least
one of them) is installed ... if so, we're good to add the
pmlogger_daily_report callback.

And this works on all platforms, independent of systemd presence and
packaging tool.