build vendored bpftool during configure
bpftool is not readily available on all systems for build and packaging.
Even on systems that do provide it, it is typically provided as a script
that points to a kernel-specific tool. Ordinarily, this matches the
provided kernel version, however inside a container the userspace
version of linux-tools available does not match the container host's
kernel version. So, instead, we build our own vendored bpftool to do
skeleton generation.