pmdabpf: set perf_buffer__poll timeout to 0
After noticing delays when sampling bpf.* metrics, we
found the multiple calls to perf_buffer__poll with a
small-number-of-milliseconds timeout were all adding
up.  After discussions with Andreas, Jason and Sohaib
we believe it is safe to set no timeout on this call,
such that it returns immediately when called (whether
values are available or not).

This change brings a 1-2sec pmdabpf sample time (with
several modules enabled) down to around 10-20msec.