pcp-dstat: add --nomissed command line option for large systems
After a number of complaints about the "N missed ticks" reporting
from dstat from users on very large systems with many disks, this
introduces a command line argument to optionally suppress it.

In the cases we've seen, its been confirmed that the increased
sample time is spent in the kernel with the large device count -
not something we can do anything about anyway.

This change has been tested by artificially inducing the failure
mode as I have no local systems showing this behaviour live.  We
cannot use archives for this either as we do not do sub-sampling
for historical data in pcp-dstat.

Resolves Red Hat BZ #2083897