selinux: rework PCP policy to use optional_policy
The way we've handled selinux policy to date has some flaws
in corner cases which continue to bite us.  One fundamental
problem is we cannot use the state of the build machine to
drive the policy we use - this can change at runtime.  I've
switch us over to using optional_policy which was designed
to handle this.

Additionally, we now ship the right bits in the right place
in order that rpm-ostree functions correctly with PCP.

Additionally, we now ship a complete ("independent") policy
for PCP based on the Fedora policy and guidelines here:
https://fedoraproject.org/wiki/SELinux/IndependentPolicy

There is a little work remaining to convert over the pcpqa
policy file, and update the tests; that'll be in a followup
commit, just wanted to get this out for initial review.  At
this stage I've also not tried to use macros throughout - I
think this would be good to do too but will require careful
auditing.

Resolves https://github.com/performancecopilot/pcp/issues/1678
Resolves Red Hat BZs #2127037, #2127037 and #1941873