packaging: fix rpm systemd-sysusers on f37 and rawhide
Recent changes to rpm/systemd have broken the unprivileged PCP
user and group account setup.  The nature of the failure is we
now see %sysuser_create_compat effects delayed until after the
PCP files are installed that need (chown) these accounts.

Resolve this by direct use of the systemd-sysusers program, at
the point where we need it to happen (%pre); happens to be the
same point we always needed it, no idea why this is delayed in
the rpm install process now - surely this will bite others. :(

This fixes CI for Fedora37 (not yet released) and rawhide.