pmdakvm: initialise perf_event interface only if/when needed
Refactor the pmdakvm startup logic so that the perf_event call
(selinux security sensitive) is done on initial access and not
during startup as before. This also makes it allocate a little
less memory in the case where no tracing metrics are accessed,
e.g. by a default zeroconf logging setup.