build: enable clang to be used
For OpenBSD, clang is the default and preferred C compiler (gcc is
there, but things like their version of Qt don't really work with
gcc on OpenBSD, which is why pmchart fell off the OpenBSD truck a
while back).

Allowing clang to be used involved some hackery in check-statics
because the compiler generates some different internal symmbol
names), and a small change in builddefs.in to push the real
compiler names into qmake so the generated Makefile has the
right commands in it.

pmchart is back in the pool on OpenBSD after these changes.