src/pmdas/ohead: new PMDA to measure overhead for groups of processes
Triggered by the on-going question "how much resource does PCP use?"
this PMDA allows us to measure this, and in general allows measurement
of resource "burn" for any group of "interesting" processes.

"Interesting" processes are identified by a regular expression match
of the "comm" (aka command) field from /proc/<pid>/stat, and then
including all the decendents of these processes ... so "pmcd" matches
pmcd and all the daemon PMDAs and any python processes running the
real PMDAs.

All metrics are dynamic.  Groups have a name, and the name is used
to populate the PMNS for some metrics and an indom for some other
metrics.

Works for long-lived processes; need sample-sleep-sample model to
compute resource "burn" so transients are missed.

No man page yet.  No help text yet.

QA in a coming commit.

This is functionally complete, but needs some polishing.