qa/common.check: small changes for _prepare_pmda() and _cleanup_pmda()
For _prepare_pmda() improve diagnostic message to make it clear
which is the agent and pmns root(s) are being operated on.

For _cleanup_pmda(), recursively remove $tmp as well as $tmp.*
for QA scripts that call _cleanup_pmda() alone in their
(exit) trap ... makes this match the default _cleanup() from
the "new" template.