rpm build: BuildRequires mandoc not man
Since man is really an alias for man-db, I can't see anything in
the man-db that the build really needs.

But we do need mandoc.

Hopefully this matches the last round of Fedora package list changes
and will put the Fedora builds back on the rails in CI ... I did not
notice it here because the man package was already installled,
independent of the package list.