build: switch to using libbpf from vendor subdir
This aims to resolve recent fallout on Fedora36 which, although
a recent Fedora, does not have the symbols we need for latest
pmdabpf modules.  As a result, the build is cleaner - we don't
need to have so much configure.ac logic for pmdabpf code.  We
can always ship all the latest modules now rather then trying
to detect newer symbols in the system libbpf and conditionally
including some.