qa/admin/list-packages+package-lists: add support for alternates
CentOS Stream9 has thrown up a case we'd not previously handled.

curl and curl-minimal are alternates ... for CI we're expecting
curl-minimal to already be installed, for my VMs curl may already
be installed and so curl-minimal cannot be installed (without using
yum --allowerasing to replace curl with curl-minimal, which is not
what I want to do).

This is now captured in the packing-list file as:

curl-minimal | curl

so that list-packages will only report curl-minimal (the primary
package) if NEITHER curl-minimal NOR curl are installed.

More than one alternate is possible.

Tags remain at the end of the line, with whitespace separation and
no preceding  "|", as in:

larry | curly | moe not4ci