zimbra PMDA: fixes to make it work with Selinux
For reasons that I don't understand (and don't want to understand),
qa/789 fails on RHEL 7.9 with Selinux Enforcing, but passes with
Selinux Permissive.  There are no AVCs in the audit log, so no policy
mumbo-jumbo will apparently fix this.

Turns out that when zimbraprobe is run from perl (the zimbra PMDA),
the Selinux context inherited from pmcd means that no variation of
su(1) or sudo(1) or runuser(1) allows the zimbaprobe shell script
(demonstrably running as root) to run zmcontrol as the user "zimbra"
(zmcontrol demands this).  We either get a Password: prompt, or System
error, or babble about not being able to post to the audit log.

When run interactively from a root shell, zimbraprobe works like
a champ.

After a series of experiments, it appears that a simple C program
(the new runaszimbra) that is run as root from zimbraprobe is able
to look up the passwd entry for the zimbra "user", extract the uid,
the home dir and the shell, and armed with this is able to:
    - setuid to the zimbra "user"
    - run zmcontrol via sh -c

The new runaszimbra is installed alongside zimbraprobe.