src/pmlogger/src: refactor getting FQDN for local host
Move DNS-sensitive calls out of the prologue and epilogue code and
do them just once before any AF registration for pmFetches, so if
there are DNS delays these do not mess with the timing of callbacks
for pmFetches in the main loop.

Should fix Nathan's qa/034 failures on a aarch64 VM with an apparent
delay in getnameinfo(2) calls.