qa/275: avoid floating point rounding non-determinism
For years (!) this test has been failing on 32-bit OpenBSD platforms
and I finally got to the bottom of it.

The derived expressions for gcov.f and gcov.g were very close to
zero, but differences in precision for doubles means the answers are
not identical.

Both expressions use a A - B style top-level expression, making this
2*A - B tests the same code paths and makes the test pass everywhere!