src/libpcp/src/accounts.c: fix buffer issue for user/group lookups
We have a bunch of routines that are wrappers (on some platforms)
for getgrgid_r(), getpwuid_r(), getpwnam_r() and getgrnam_r()) ...
in all cases, we declare a namebuf[] on the stack sized by some
arbitrary number (1024 for the most part, sometimes 512).

This completely ignores the sysconf() API that is there to provide
guidance on the correct buffer sizes!

And of course, this eventually broke (on OpenBSD, where namebuf[]
needs to be 2600+ not 1024 in one case).

Replace the on-stack buffers by a single buffer that is sized to be
the maximum of all the various sysconf() suggestions ... the buffer is
only used for transient data, so can be initialized in some one-trip
code and re-used thereafer.  This does add a couple more statics,
so some check-statics changes are needed as well.