pmlogrewrite: add missing code to remove unreferenced indom
If an indom appears in the input archive, but after rewriting
there is no reference to the indom, then the indom should
not appear in the metadata of the output archive.

This commit fixes https://github.com/performancecopilot/pcp/issues/1743

Also qa/822 is now a real test and no longer reserved.

And qa/1404 is a new valgrind version of qa/822.

And qa/488.out needed to be remade.