src/pmlogrewrite/pmlogrewrite.c & qa: more indom culling changes
Rework the second part of the fix for
https://github.com/performancecopilot/pcp/issues/1743 specifically
dealing with cases where the indom id is renumbered.  The new
implementation uses more robust reference counting for the _input_
archive indom, rather than the earlier weaker reference counting and
a heuristic that worked for qa/822 but caused other qa fallout.

Also cleaned up a bunch of minor qa fluff related to pmlogrewrite
testing.