rc scripts: ensure pmcd.conf not rewritten unnecessarily
It was observed that pmcd.conf inode modification times were
updated on every restart of pmcd.  Turns out its because of
the PMDA Upgrade scripts, where some were using 'sed -i' to
affect a no-op change to the contents.  We're now take more
care ensuring sed is only done when the contents of the file
are definitely going to be modified.

Resolves Red Hat BZ #2166819.