build: check consistency for error codes
We define the PM_ERR_* error codes in multiple places:
- pmapi.h
- err.c in libpcp
- (2 places in) PMDA.pm for Perl
- pmapi.c for Python

The src/libpcp/src/check-errorcodes script checks these are consistent
and is used in the libpcp build.

A couple of cosmetic "fixups" are included (the script checks the
string used in pmErrStr() against the description in comments in a
couple of places).

No issue lurking here at the moment ... this is defensive and triggered
by some unrelated work that was exploring changes that would have
required a new error code.