Introducing PMCD_HOSTNAME_CHANGE
We've come across some situations where a host may change hostname,
either during post-boot activity when transitioning from a temporary
to a permanent hostname, or by subsequent administrative action.

In either case, we ended up with a disconnect for the primary pmlogger
in particular where it appeared to be creating archives in a directory
that matched an old hostname, and not the current hostname.  Worse,
because this pmlogger does not match any entry in the control files
(after LOCALHOSTNAME is translated to the current hostname), the
periodic and nightly archive maintenance (compression, rotation,
merging and culling) were not being done correctly.

The solution is to use the existing mechanism for returning bit-fields
of status from pmcd via from pmFetch() to indicate a change in
pmcd's state.  Specifically we've added a new PMCD_HOSTNAME_CHANGE bit.

Clients that care can test for this and take appproriate action.

In the case of pmlogger, it will exit if it was launched (or re-exec'd)
from pmlogger_check or pmlogger_daily.  This will allow the systemd
or cron mechanisms to relaunch the pmlogger using the current hostname
against the control files.

qa/1441 (new) tests this facility.

This commit addresses some of the issues raised in
https://bugzilla.redhat.com/show_bug.cgi?id=2172892
... still to be done is the janitorial work for archives created
by a pmlogger does not match the current state of the control files
and hostname(1).