qa/1441 (and 1443): combine tests
Doing the change hostname thing is sufficiently disruptive that
it is better to do it once and check both pmlogger and pmie.
(this does assume we have hosts with a primary pmie enabled
or the pmlogger part of the test will be _notrun ... but this
does not run in CI and in the QA Farm there are ample hosts
running a primary pmie).

Also 1441 introduced the possibility of mucking with pmie as
a by-product, and this was not checked for nor cleaned up (more
precisely waited for), so 1443 could crash into the debris from 1441.

For both reasons, 1443 has been merged into 1441 and 1443 removed.