pmie: Fix another code path that is exposed to pmcd hostname changes
If pmie has lost connection to pmcd, then it stays in a state where
it attempts pmReconnectContext() with back-off until pmcd comes back
to life.  But if the hostname for pmcd has changed between the last
pmFetch() and now, we never see a PMCD_HOSTNAME_CHANGE state change.

Fixing this involves some extra logic after pmReconnectContext()
succeeds, and a call to the (extra handy) pmGetContextHostName()
and conditionally forcing an exit if the hostname has changed.

Because PCP's QA bounces pmcd (a lot!) and pmie will continue to
run through much of this, qa/1441 was sometimes seeing a pmie in
the pmReconnectContext() back-off state, and the hostname change was
being missed.