pmlogger_janitor: new script called from pmlogger_check
- kills off any running pmlogger that was started from pmlogger_check
  (or pmlogger_daily or pmlogctl) that no longer has a matching entry
  in the pmlogger control files (/etc/pcp/pmlogger/control and
  the files in /etc/pcp/pmlogger/control.d)
- for any pmlogger killed off above, go find the directory it was
  writing archives into, and compress all the archive files there

The only pmloggers impacted by this will be ones that were launched
from control files, but because the control files have changed or the
hostname has changed (and the earlier pmcd+pmlogger changes did not
catch the hostname change) they are not longer manageable by pmlogctl,
pmlogger_check and pmlogger_daily ... in other words they are run
away orphans.

Because pmlogger_janitor is only ever intended to be called from
pmlogger_check, there is no man page for pmlogger_janitor.