libpcp: add pmAddDerivedText()
A new API to allow help text (one line and multi-line) to be associated
with derived metrics.

Extends the capability that was previously only available through the
"oneline" and "helptext" attribute mechanism for pmLoadDerivedConfig().

Also fixed on the pmLookupText() code path so that it works for both
global and per-context derived metrics.

Fixed a lock inversion problem in __dmtraverse() that I noticed while
driving by.

New QA (1445 and 1446) and man page for pmAddDerivedText(3).