pmieconf: add support for a action testing rule
When testing and documenting pmieconf actions like the new
webhook endpoint action it's extremely handy to have a test
rule that we can use to verify the setup.  This needs to be
in pmieconf format and on relativly short interval, without
a holdoff, in order for it to be quickly observable.