src/pmlogger/utilproc.sh: refactor _do_dir_and_args()
In the related pmlogger -d work, a number of problems were identified
in _do_dir_and_args() which is responsible for mangling, cuting and
pasting together the real "directory" field from a pmlogger control
file ... this is "hard" because the control file has whitespace
separated fields and any `...` and $(...) constructs in the directory
field will also typically include whitespace.

- fix whitespace issues in re-assembly
- implement single quote (') enclosing the whole directory ... same
  as double quote (") and as documented in pmlogger.control(5), but
  NOT implemented previously
- cleanly separate "inside a quote" processing from shell expansion
  processing
- handle single word (no whitespace) constructs for `...` and $(...)
  without requiring quoting
- add (secret) debugging via $debug_do_dir_and_args which can be set
  to "true" in QA scripts

Also updates and remakes for qa/1094 and qa/1425.