pmlogger: add -d directory argument
The "directory" may be literal or a template (to be expanded at
run-time by sh(1)).

This provides a means for pmlogger_check to pass the directory
field from a control file to pmlogger, then have pmlogger expand
this each time it is started or re-exec'd.

This commit includes the pmlogger changes, man page updates and
shell completion changes for bash(1) and zsh(1).