src/pmlogger/pmlogger_check.sh: pass directory field to pmlogger -d
These changes re-instate the semantics for the pmlogger control files
from before pcp-5.1.0, namely shell metacharacters in the directory
field are evaluated each time it is started or re-exec'd (the latter
case was lost when re-exec was introduced).

This change fixes https://github.com/performancecopilot/pcp/issues/1753
and provides a permanent fix for the issues raised by Jeff Avila in
respect of the SUPReMM module's instructions for configuring PCP, see
https://supremm.xdmod.org/10.0/supremm-compute-pcp.html, where the
directory field of the control file looks like:
"/gpfs/data/root/pcp/$(date +%Y)/$(date +%m)/LOCALHOSTNAME/$(date +%Y)-$(date +%m)-$(date +%d)"

The handling of arguments in pmlogger_check is complicated and the
process linking between pmlogger_check and systemd (when the latter
is in the mix) is obscure and brittle in the extreme.  Consequently
these changes pass the directory field from the control file to
pmlogger -d with the following rules:
- the field is enclosed in double quotes
- all spaces are mapped to CTL-A
- all tabs are mapped to CTL-B

pmlogger knows about these rules and reconstructs the original
directory field before using it.

Also in this commit are some loosely related changes and fixes:
- -l /dev/tty will output all messages from pmlogger_check to stdout
  ... no hiding them in some logfile, no .prev logfile juggling, no
  exec to juggle output streams
- use $orig_dir from _do_dir_and_args() to set the -d option for
  pmlogger
- handle spaces in the expanded directory field
- report early diagnostics from pmlogger ... some of these happen
  before the -l logfile is opened and were being lost (which is
  unfortunate because some of them are fatal and pmlogger vanishes
  without a trace!)