src/pmlogger/utilproc.sh: fix for OpenBSD
In something from the "stranger than fiction" chest comes this gem.

On OpenBSD 6.0.4 (vm37)
	$ xv -dc ... | dd ibs=1 count=7
HANGS the xv process when the dd eats 7 bytes and exits.

Special case for OpenBSD to avoid the pipeline and use a temportart
file to decompress the whole file.  This sucks, but it does not hang.