qa/common.rc: dodge OpenBSD bullet with /var/tmp
On OpenBSD 6.0.4 (vm37), /var/tmp is a symlink to ../tmp.

So /var/tmp/foo and /tmp/foo are different pathnames for EXACTLY the
same file.

This was confusing some of the pmlogger_check QA tests in particular
where the archives were being created below $tmp, and the attempt
to matchup loggers was failing because the shell scripts saw
the /var/tmp/foo name, but pmlogger uses getcwd() to generate the
/tmp/foo name for the /var/lib/pcp/pmlogger/XXXX port map files and
pmlogger_check totally lost the plot because the directory matching
logic failed.

For all practical purposes, the problem is confined to QA, so just
be a bit more careful here when setting $tmp, and use the realpath(1)
variant of /var/tmp as the directory prefix.