common.check: fix small race in _wait_pmlogger_end()
The race is the window between the [ ... ] guard and the
    ls -l "$PCP_TMP_DIR/pmlogger/$__pid

If the file gets removed in this window (and that's not unexpected)
then the ls fails and babbles a warning on stderr that ends up in
$seq.out.bad.

Simply redirect stderr into the $seq.full along with stdout from the
ls command.

Observed on vm33 (OpenBSD 7.2) with qa/061, although the problem
potentially exists for multiple QA tests on any platform.  But it is
inherently non-deterministic, which may expalain why no analysis has
found the problem previously.