src/pmlogextract/pmlogextract.c: improve resilence in QA environment
pmlogger needs several round-trips to pmcd when a new metric is
encountered in a pmResult to get:
    - pmDesc
    - PMNS name(s)
    - InDom (if it is a new InDom or looks like it has changed)

If the associated PMDA is removed in the middle of this, or (in the
case of a PMDA with dynamic metrics like the mmv PMDA) the metric
simply goes away, we end up with a mangled archive with a pmResult
and incomplete metadata.

pmlogger is at fault here (I'll look to fix that later, but it is going
to be messy), but unfortunately pmlogcheck does not spot the problem
(another thing to fix subsequently), pmlogrewrite cannot repair the
archive because it does not have the metadata to drive its rewrite
rules (this can't be fixed), and pmlogextract spits the dummy.

This commit addresses the pmlogextract dummy spit, by removing the
offending metric from the pmResult (it is useless, so no loss).

For all practical considerations, the problem is confined to hosts
running QA tests as most production environments do not have PMDAs
and metrics that come and go like we see during QA runs.

Also included here is an example bad archive and a new QA test
(1449) to test the pmlogextract fix.