qa/816: extend and remake after recent pmdumptext change
This test exposed one of the half-baked corner cases where pmdumptext
might, or might not, perform canonical scaling of values (based on
units like Kbyte or Mbyte or ...) ... this required a remake.

The values with -c config were (before the last pmdumptext commit)
different to the values reported with the same metricspec on the
command line, so extend the test to check for this.