pmie: ensure pmie.log.YYYYMMDD is owned by pcp:pcp
Fixes the issue that is raised in the circa June 2023 comments
in https://bugzilla.redhat.com/show_bug.cgi?id=1832262.

Add new qa test (1453) that reproduced the problem and shows it
to be fixed.

What remains as something of a mystery is why when this file is
created by pmie_check the problem is observed, but when the file
is created by pmie_daily the problem does not happen.  A chown
has been added to *both* scripts as an Irish fix (to be sure, to
be sure).