src/pmdas/linux/getinfo.c: fix SEGV when lsb-release is the only /etc/*-release file
Bug in get_distro_info() as triaged and fixed by Tim Guo(BJ-RD)
<TimGuo@zhaoxin.com>.

Ick (as per comment already in code) logic for lsb-release was null
terminating the array *after* dinking with the base address of the
array ... do the null termination first and the problem goes away.

Most of us have never seen this because at least one of the other
*-release files is found in /etc and we never get the the lsb-release
case.