pmdaopenmetrics: reduce transient instance logs, improve labels
Ensure an indicative error code is used that does not result
in openmetrics.log getting a diagnostic every time instances
vanish.  This is common with process_exporter, for example.

Perform less processing of quoting in labels so that we are
able to handle metric from the systemd_exporter sensibly and
without parse errors in libpcp (partially stripped escaping
is the root cause there).  Drop per-instance optional labels
as well as this just results in unhelpful duplicate text.

Finally, update QA to match, adding in a problematic systemd
case to the openmetrics regression samples.