man pages: retire "archive log" phrase
We've been troubled by this for a long time ... the phrase "archive log"
is both needlessly wordy and confusing with "log" to mean something like
like pmcd.log or pmlogger.log.

    [PCP] archive log(s) -> [PCP] archive(s)
    log -> archive (where this is referring to an archive, not a log)

We're stuck with "log" as a verb and LOG in macros and Log in PMAPI
routine names, but that's not going to change in my lifetime.